+++ extensions/zed/languages/prosaic/highlights.scm
@@ -1,0 +1,8 @@
+(comment) @comment
+(annotation) @comment.doc
+(control_keyword) @keyword
+(action_verb) @function
+(block_label) @label
+(file_path) @string.special
+(template_var) @string.escape
+(operator) @operator